report bug when docker fail

+++ controler/container_rm.py
@@ -25,7 +25,8 @@ def container_rm(db:OpenGaussConnector, cid:int, uid:int):
     # 删除容器
     docker_cmd = f'docker -H ssh://root@{ip} rm c{cid}'
     out = subprocess.run(docker_cmd, capture_output=True, shell=True)
-    
+    if out.returncode != 0:
+        return -1, f'docker failed with message {out.stderr.decode('utf-8')}'
     # 更新container表
     cmd = f'delete from containers where cid = {cid}'
     db.exec(cmd)
